+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Versioning].
 
 ## Unreleased
 
-- _No changes yet_
+- ([#2054]) Allow top-level side-effects in scripts.
 
 ## 3.6.3 (2026-06-25)
 
@@ -273,3 +273,4 @@ Versioning].
 [#1879]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1879
 [#1884]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1884
 [#1947]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1947
+[#2054]: https://github.com/ericcornelissen/eslint-plugin-top/pull/2054